Move interface into separate source file
